Add tracing from http handlers
