+++ .gitignore
@@ -1,2 +1,1 @@
 # Go binary
-payforadoption-go
